@@ -15,6 +15,7 @@
     "react": "19.1.0",
     "react-dom": "19.1.0",
     "react-icons": "^5.5.0",
+    "recharts": "^2.12.7",
     "tailwind-merge": "^3.3.1"
   },
   "devDependencies": {
@@ -32,6 +33,6 @@
     "jest": "^29.7.0",
     "tailwindcss": "^4",
     "ts-jest": "^29.1.4",
-    "typescript": "^5"
+    "typescript": "5.9.2"
   }
 }
