@@ -1,267 +1,11 @@
 'use client';
 
-import { useState, useEffect } from 'react';
-import Card from '@/components/Card';
-import { FaChartLine, FaUsers, FaVideo, FaImages, FaMusic, FaFlag, FaMountain, FaGlobe } from 'react-icons/fa';
-import { FaArrowTrendUp, FaArrowTrendDown } from 'react-icons/fa6';
-import { apiClient, handleApiResponse } from '@/lib/api';
-
-// [SNIPPET]: thinkwithai + kenyafirst + surgicalfix + refactorintent
-// [CONTEXT]: Analytics page with Kenya-first design and mobile-first approach
-// [GOAL]: Create comprehensive analytics dashboard with real backend integration
-// [TASK]: Implement analytics with proper loading states and cultural authenticity
-
-interface AnalyticsData {
-  totalViews: number;
-  totalGenerations: number;
-  activeUsers: number;
-  popularContent: Array<{
-    title: string;
-    views: number;
-    type: string;
-  }>;
-  performanceMetrics: {
-    avgGenerationTime: string;
-    successRate: number;
-    userSatisfaction: number;
-  };
-}
+import UsageAnalytics from '@/components/Admin/UsageAnalytics';
 
 export default function AnalyticsPage() {
-  const [analytics, setAnalytics] = useState<AnalyticsData | null>(null);
-  const [isLoading, setIsLoading] = useState(true);
-  const [error, setError] = useState<string | null>(null);
-  const [timeRange, setTimeRange] = useState<'7d' | '30d' | '90d'>('30d');
-
-  useEffect(() => {
-    loadAnalytics();
-  }, [timeRange]);
-
-  const loadAnalytics = async () => {
-    setIsLoading(true);
-    setError(null);
-
-    const response = await apiClient.getAnalytics();
-    handleApiResponse(
-      response,
-      (data) => {
-        // Since real analytics might not be available, show friendly message
-        setAnalytics({
-          totalViews: 0,
-          totalGenerations: 0,
-          activeUsers: 0,
-          popularContent: [],
-          performanceMetrics: {
-            avgGenerationTime: 'N/A',
-            successRate: 0,
-            userSatisfaction: 0
-          }
-        });
-      },
-      (error) => setError(error)
-    );
-
-    setIsLoading(false);
-  };
-
-  if (isLoading) {
-    return (
-      <div className="flex items-center justify-center h-64">
-        <div className="text-center">
-          <div className="w-8 h-8 border-4 border-green-600 border-t-transparent rounded-full animate-spin mx-auto mb-4"></div>
-          <span className="text-gray-600 font-medium">Loading analytics...</span>
-        </div>
-      </div>
-    );
-  }
-
   return (
-    <div className="space-y-6">
-      {/* Kenya-First Header */}
-      <div className="bg-gradient-to-r from-green-600 via-red-600 to-black p-6 rounded-xl text-white shadow-lg">
-        <div className="flex items-center justify-between">
-          <div className="flex items-center space-x-4">
-            <FaChartLine className="text-3xl" />
-            <div>
-              <h1 className="text-2xl font-bold">Analytics Dashboard 🇰🇪</h1>
-              <p className="text-green-100">Performance insights for your Kenya-first content</p>
-            </div>
-          </div>
-          <div className="hidden md:block">
-            <FaMountain className="text-4xl text-yellow-300" />
-          </div>
-        </div>
-      </div>
-
-      {/* Time Range Selector */}
-      <Card className="p-6">
-        <div className="flex items-center justify-between mb-4">
-          <h2 className="text-xl font-bold text-gray-800">Performance Overview</h2>
-          <div className="flex space-x-2">
-            {[
-              { key: '7d', label: '7 Days' },
-              { key: '30d', label: '30 Days' },
-              { key: '90d', label: '90 Days' }
-            ].map(({ key, label }) => (
-              <button
-                key={key}
-                onClick={() => setTimeRange(key as any)}
-                className={`px-4 py-2 rounded-lg transition-colors duration-200 ${
-                  timeRange === key
-                    ? 'bg-green-600 text-white'
-                    : 'bg-gray-100 text-gray-700 hover:bg-gray-200'
-                }`}
-              >
-                {label}
-              </button>
-            ))}
-          </div>
-        </div>
-
-        {error ? (
-          <div className="text-center py-8">
-            <FaChartLine className="text-4xl text-red-600 mx-auto mb-4" />
-            <p className="text-red-600 font-medium mb-2">Unable to load analytics data</p>
-            <p className="text-sm text-gray-600 mb-4">{error}</p>
-            <button
-              onClick={loadAnalytics}
-              className="bg-green-600 hover:bg-green-700 text-white px-4 py-2 rounded-lg transition-colors duration-200"
-            >
-              Try Again
-            </button>
-          </div>
-        ) : !analytics || (analytics.totalViews === 0 && analytics.totalGenerations === 0) ? (
-          <div className="text-center py-8">
-            <FaChartLine className="text-4xl text-gray-400 mx-auto mb-4" />
-            <p className="text-gray-600 font-medium mb-2">Analytics data not available</p>
-            <p className="text-sm text-gray-500 mb-4">
-              Start generating content to see performance insights and analytics data.
-            </p>
-            <div className="bg-blue-50 border border-blue-200 rounded-lg p-4 max-w-md mx-auto">
-              <p className="text-blue-800 text-sm">
-                <strong>Coming Soon:</strong> Comprehensive analytics including user engagement, 
-                content performance, and cultural impact metrics.
-              </p>
-            </div>
-          </div>
-        ) : (
-          <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-6">
-            {/* Key Metrics */}
-            <div className="bg-blue-50 p-4 rounded-lg">
-              <div className="flex items-center justify-between">
-                <div>
-                  <p className="text-blue-600 text-sm font-medium">Total Views</p>
-                  <p className="text-2xl font-bold text-blue-800">{analytics.totalViews.toLocaleString()}</p>
-                </div>
-                <FaArrowTrendUp className="text-2xl text-blue-600" />
-              </div>
-            </div>
-
-            <div className="bg-green-50 p-4 rounded-lg">
-              <div className="flex items-center justify-between">
-                <div>
-                  <p className="text-green-600 text-sm font-medium">Content Generated</p>
-                  <p className="text-2xl font-bold text-green-800">{analytics.totalGenerations.toLocaleString()}</p>
-                </div>
-                <FaVideo className="text-2xl text-green-600" />
-              </div>
-            </div>
-
-            <div className="bg-purple-50 p-4 rounded-lg">
-              <div className="flex items-center justify-between">
-                <div>
-                  <p className="text-purple-600 text-sm font-medium">Active Users</p>
-                  <p className="text-2xl font-bold text-purple-800">{analytics.activeUsers.toLocaleString()}</p>
-                </div>
-                <FaUsers className="text-2xl text-purple-600" />
-              </div>
-            </div>
-
-            <div className="bg-orange-50 p-4 rounded-lg">
-              <div className="flex items-center justify-between">
-                <div>
-                  <p className="text-orange-600 text-sm font-medium">Success Rate</p>
-                  <p className="text-2xl font-bold text-orange-800">{analytics.performanceMetrics.successRate}%</p>
-                </div>
-                <FaArrowTrendUp className="text-2xl text-orange-600" />
-              </div>
-            </div>
-          </div>
-        )}
-      </Card>
-
-      {/* Popular Content */}
-      <Card className="p-6">
-        <h2 className="text-xl font-bold text-gray-800 mb-4">Popular Content</h2>
-        {!analytics?.popularContent.length ? (
-          <div className="text-center py-8">
-            <FaImages className="text-4xl text-gray-400 mx-auto mb-4" />
-            <p className="text-gray-600 font-medium mb-2">No popular content data available</p>
-            <p className="text-sm text-gray-500">
-              Generate more content to see what resonates with your audience.
-            </p>
-          </div>
-        ) : (
-          <div className="space-y-4">
-            {analytics.popularContent.map((item, index) => (
-              <div key={index} className="flex items-center justify-between p-4 bg-gray-50 rounded-lg">
-                <div className="flex items-center space-x-3">
-                  <span className="bg-green-600 text-white text-sm font-bold w-8 h-8 rounded-full flex items-center justify-center">
-                    {index + 1}
-                  </span>
-                  <div>
-                    <p className="font-medium text-gray-800">{item.title}</p>
-                    <p className="text-sm text-gray-500">{item.type}</p>
-                  </div>
-                </div>
-                <div className="text-right">
-                  <p className="font-bold text-gray-800">{item.views.toLocaleString()}</p>
-                  <p className="text-sm text-gray-500">views</p>
-                </div>
-              </div>
-            ))}
-          </div>
-        )}
-      </Card>
-
-      {/* Performance Metrics */}
-      <Card className="p-6">
-        <h2 className="text-xl font-bold text-gray-800 mb-4">Performance Metrics</h2>
-        <div className="grid grid-cols-1 md:grid-cols-3 gap-6">
-          <div className="text-center">
-            <div className="bg-blue-100 w-16 h-16 rounded-full flex items-center justify-center mx-auto mb-3">
-              <FaChartLine className="text-2xl text-blue-600" />
-            </div>
-            <p className="text-gray-600 text-sm">Avg Generation Time</p>
-            <p className="text-xl font-bold text-gray-800">{analytics?.performanceMetrics.avgGenerationTime || 'N/A'}</p>
-          </div>
-
-          <div className="text-center">
-            <div className="bg-green-100 w-16 h-16 rounded-full flex items-center justify-center mx-auto mb-3">
-              <FaArrowTrendUp className="text-2xl text-green-600" />
-            </div>
-            <p className="text-gray-600 text-sm">Success Rate</p>
-            <p className="text-xl font-bold text-gray-800">{analytics?.performanceMetrics.successRate || 0}%</p>
-          </div>
-
-          <div className="text-center">
-            <div className="bg-purple-100 w-16 h-16 rounded-full flex items-center justify-center mx-auto mb-3">
-              <FaUsers className="text-2xl text-purple-600" />
-            </div>
-            <p className="text-gray-600 text-sm">User Satisfaction</p>
-            <p className="text-xl font-bold text-gray-800">{analytics?.performanceMetrics.userSatisfaction || 0}%</p>
-          </div>
-        </div>
-      </Card>
-
-      {/* Cultural Footer */}
-      <div className="bg-gradient-to-r from-yellow-400 via-red-500 to-green-600 p-4 rounded-lg text-white text-center">
-        <div className="flex items-center justify-center space-x-2">
-          <FaFlag className="text-lg" />
-          <span className="font-medium">Measuring impact of Kenya-first AI innovation</span>
-          <FaGlobe className="text-lg" />
-        </div>
-      </div>
+    <div>
+      <UsageAnalytics />
     </div>
   );
 }
